Removed obsolete IExitOperation interface and its implementations, as they have been merged into single ListExitOperation class. Resolved various Resharper warnings.

+++ src/NHibernate.Shards/Cfg/IShardConfiguration.cs
@@ -18,7 +18,7 @@ namespace NHibernate.Shards.Cfg
         /// <summary>
         /// unique id of the shard
         /// </summary>
-        int ShardId { get; }
+        short ShardId { get; }
 
         /// <summary>
         /// the cache region prefix for the shard
